refactor(types): refactoring documentarray and memmap via mixin
