+++ tests/unit/types/arrays/documentarray/test_matcharray.py
