+++ tests/unit/types/arrays/memmap/test_memmap.py
@@ -357,7 +357,7 @@ def test_memmap_save_reload(tmpdir):
     for doc in dam1:
         assert doc.content == 'hello world'
 
-    dam.save()
+    dam.flush()
     dam1.reload()
 
     # dam from disk
@@ -509,7 +509,7 @@ def test_embeddings_wrong_len(tmpdir):
     dam.extend([Document() for x in range(100)])
     embeddings = np.ones((2, 10, 10))
 
-    with pytest.raises(ValueError, match='the number of rows in the'):
+    with pytest.raises(ValueError):
         dam.embeddings = embeddings
 
 
@@ -529,33 +529,6 @@ def test_blobs_setter_dam(tmpdir):
     np.testing.assert_almost_equal(dam.blobs, blobs)
     for x, doc in zip(blobs, dam):
         np.testing.assert_almost_equal(x, doc.blob)
-
-
-def test_tags_getter_dam(tmpdir):
-    dam = DocumentArrayMemmap(tmpdir)
-    dam.extend([Document(tags={'a': 2, 'c': 'd'}) for _ in range(100)])
-    assert len(dam.tags) == 100
-    assert dam.tags == dam.get_attributes('tags')
-
-
-def test_tags_setter_dam(tmpdir):
-    dam = DocumentArrayMemmap(tmpdir)
-    tags = [{'a': 2, 'c': 'd'} for _ in range(100)]
-    dam.extend([Document() for _ in range(100)])
-    dam.tags = tags
-    assert dam.tags == tags
-
-    for x, doc in zip(tags, dam):
-        assert x == doc.tags
-
-
-def test_setter_wrong_len(tmpdir):
-    dam = DocumentArrayMemmap(tmpdir)
-    dam.extend([Document() for _ in range(100)])
-    tags = [{'1': 2}]
-
-    with pytest.raises(ValueError, match='the number of tags in the'):
-        dam.tags = tags
 
 
 def test_texts_getter_dam(tmpdir):
@@ -583,7 +556,7 @@ def test_texts_wrong_len(tmpdir):
     dam.extend([Document() for _ in range(100)])
     texts = ['hello']
 
-    with pytest.raises(ValueError, match='the number of texts in the'):
+    with pytest.raises(ValueError):
         dam.texts = texts
 
 
@@ -592,7 +565,7 @@ def test_blobs_wrong_len(tmpdir):
     dam.extend([Document() for x in range(100)])
     blobs = np.ones((2, 10, 10))
 
-    with pytest.raises(ValueError, match='the number of rows in the'):
+    with pytest.raises(ValueError):
         dam.blobs = blobs
